requirements/production.txt
Django
psycopg2
prettyconf
dj-database-url
djangorestframework
djangorestframework-recursive
tokenlib
rest_framework_swagger